Module forest::paramfetch

Structs

FetchProgress

ParameterData

Enums

SectorSizeOpt
Sector size options for fetching.

Constants

DEFAULT_PARAMETERS

DIR_ENV

GATEWAY

GATEWAY_ENV

PARAM_DIR

TRUST_PARAMS_ENV

Functions

check_file

fetch_params

fetch_verify_params

get_params
Get proofs parameters and all verification keys for a given sector size given a param JSON manifest.

get_params_default
Get proofs parameters and all verification keys for a given sector size using default manifest.

param_dir

Type Definitions

ParameterMap