Module forest::cli

Re-exports

pub use self::config::Config;

Modules

auth_cmd

chain_cmd

config

fetch_params_cmd

genesis_cmd

mpool_cmd

net_cmd

state_cmd

sync_cmd

wallet_cmd

Structs

Cli

CLI structure generated when interacting with Forest binary

CliOpts

CLI options

Enums

Subcommand

Forest binary subcommands available.

Functions

balance_to_fil

Convert an atto FIL balance to FIL

block_until_sigint

Blocks current thread until ctrl-c is received

cli_error_and_die

Print an error message and exit the program with an error code Used for handling high level errors such as invalid params

format_vec_pretty

Format a vector to a prettified string

handle_rpc_err

Print a stringified JSON-RPC error and exit

print_rpc_res

Prints a plain HTTP JSON-RPC response result

print_rpc_res_bytes

Prints a bytes HTTP JSON-RPC response result

print_rpc_res_cids

Prints a tipset from a HTTP JSON-RPC response result

print_rpc_res_pretty

Prints a pretty HTTP JSON-RPC response result

print_stdout

Prints a string HTTP JSON-RPC response result to a buffered stdout

to_size_string

convert bigint to size string using byte size units (ie KiB, GiB, PiB, etc)